move RunsCounter to separate module and add a test
